fixed bug, was missing a third HOMEPAGE_URL
